修改模板编译,增加layout布局

+++ Framework/View.php
@@ -31,17 +31,41 @@ class View
         $this->parseTemplate($content);
     }
 
-    public function parseTemplate($content)
+    /**
+     * @param $content
+     */
+    public function replaceTemplate($content)
     {
         preg_match_all('#<include\s+file=\s?(\"|\')\s?(.*?)\s?(\"|\')\s+/>#', $content, $arr);
-        foreach ($arr[2] as $k => $v) {
-            $s_include = file_get_contents(APP_PATH . 'View' . DS . Module_GROUP . DS . 'Base' . DS . $v);
-            $content  = str_replace($arr[0][$k], $s_include, $content);
-        }
-        $cache_file = APP_PATH. 'Cache/'.md5(Module_CONTROLLER.DS.Module_ACTION).'.php';
-        $content = preg_replace('#\{\$([a-zA-Z]+)\}#', '<?php echo \$this->value[\'\\1\']; ?>', $content);
+        if ($arr) {
+            foreach ($arr[2] as $k => $v) {
+                $s_include = file_get_contents(APP_PATH . 'View' . DS . Module_GROUP . DS . 'Base' . DS . $v);
+                $content = str_replace($arr[0][$k], $s_include, $content);
 
-        file_put_contents($cache_file, $content);
+            }
+        }
+        return $content;
+    }
+
+    public function parseTemplate($content)
+    {
+        if (C('LAYOUT') && ( strpos($content,'__NO_LAYOUT__')===false)) {
+            $layout = file_get_contents(APP_PATH . 'View' . DS . Module_GROUP . DS . 'layout.html');
+            $content = str_replace('__CONTENT__', $content, $layout);
+        } else {
+            $content = str_replace('__NO_LAYOUT__', '', $content);
+        }
+
+        $content = $this->replaceTemplate($content);
+        /*if (preg_match('#\{\$([a-zA-Z]+)\}#', $content)) {
+            $content = preg_replace('#\{\$([a-zA-Z]+)\}#', '<?php echo \$this->value[\'\\1\']; ?>', $content);
+       } */
+
+        $cache_file = APP_PATH. 'Cache'.DS.Module_GROUP.DS.md5($content).'.php';
+
+        if (!file_exists($cache_file)) {
+            file_put_contents($cache_file, $content);
+        }
 
         include $cache_file;
     }
